refactor: split impl into apis
