+++ lib/client-base.js
@@ -666,7 +666,7 @@ function resume (client) {
     }
 
     const request = client[kQueue][client[kPendingIdx]]
-    if (request.finished) {
+    if (request.aborted) {
       // Request was aborted.
       // TODO: Avoid splice one by one.
       client[kQueue].splice(client[kPendingIdx], 1)
